Move `AggregateStoreRepository` tests into `providertest`.
